chore: moved themeMerge's export to the utils folder, updated rollup config
