+++ .github/OWNERS
@@ -39,6 +39,6 @@
 /src/Form/ControlLabel @brandonlilly
 /src/Table @brandonlilly
 /src/Tabs @brandonlilly
-/src/theme/themeMerge @zero0Halo
+/src/utils/themeMerge @zero0Halo
 /src/Skeleton @brandonlilly
 /src/Badge @bjohnsonrmn
