+++ src/index.ts
@@ -23,7 +23,7 @@ export { Menu, Item as MenuItem } from './Menu';
 export { PopOver } from './PopOver';
 export { Typography } from './Typography';
 export { Tooltip } from './Tooltip';
-export { RootTheme, themeMerge, ThemeProvider } from './theme';
+export { RootTheme, ThemeProvider } from './theme';
 export { Card } from './Card';
 export { Hero } from './Hero';
 export { Tabs } from './Tabs';
@@ -31,4 +31,5 @@ export { Skeleton } from './Skeleton';
 export { Pagination } from './Pagination';
 export { Modal, BaseModalBackground, ModalProvider } from './Modal';
 export { Input, Toggle, Radio, Checkbox, ControlLabel } from './Form';
+export { themeMerge } from './utils';
 export * from './Icon';
